Feat: changed to handle map and passport
